Loading incidents into IncidentService
